+++ src/Session.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "MsgPackRpc.hpp"
-#include "Timer.hpp"
 #include "Renderer.hpp"
 #include "RedrawHandler.hpp"
 #include "Input.hpp"
@@ -31,11 +30,15 @@ public:
         return !_nvim_exited.load(std::memory_order_relaxed);
     }
 
+    const std::string& GetOutput() const
+    {
+        return _rpc->GetOutput();
+    }
+
 private:
     uv_pipe_t _stdin_pipe, _stdout_pipe;
     uv_process_t _child_req;
     std::unique_ptr<MsgPackRpc> _rpc;
-    std::unique_ptr<Timer> _timer;
     std::unique_ptr<Renderer> _renderer;
     std::unique_ptr<RedrawHandler> _redraw_handler;
     std::unique_ptr<Input> _input;