+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 
 - Switched to gtk4 to get rid of texture noise. Chunks of texts are rendered as gtk labels.
+- Capture and display service output like --help or --version in the GUI window
 
 ### Fixed